remove soundtouch, fixes for Snow Leopard build, use libintl explicitly on non-glibc systems, fix for new glib include rules
git-svn-id: svn://localhost/ardour2/branches/2.0-ongoing@12031 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ gtk2_ardour/engine_dialog.cc
@@ -668,7 +668,7 @@ EngineControl::enumerate_coreaudio_devices ()
 	// (code snippet gently "borrowed" from St?hane Letz jackdmp;)
 	OSStatus err;
 	Boolean isWritable;
-	size_t outSize = sizeof(isWritable);
+	UInt32 outSize = sizeof(isWritable);
 
 	backend_devs.clear ();
 
@@ -684,8 +684,8 @@ EngineControl::enumerate_coreaudio_devices ()
 		if (err == noErr) {
 			// Look for the CoreAudio device name...
 			char coreDeviceName[256];
-			size_t nameSize;
-
+			UInt32 nameSize;
+			
 			for (int i = 0; i < numCoreDevices; i++) {
 
 				nameSize = sizeof (coreDeviceName);
